Feature/fix issues

+++ src/main/java/com/kraken/api/model/Invoice.java
@@ -1,15 +1,14 @@
 package com.kraken.api.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.kraken.api.web.rest.v1.validator.annotations.ValidDateTime;
-import jakarta.annotation.Nullable;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Builder
@@ -20,11 +19,11 @@ public record Invoice(
         @NotBlank(message = "invoiceNumber must not be blank")
         String invoiceNumber,
         @NotNull(message = "grossAmount must not be empty")
-        Double grossAmount,
+        BigDecimal grossAmount,
         @NotNull(message = "gstAmount must not be empty")
-        Double gstAmount,
+        BigDecimal gstAmount,
         @NotNull(message = "netAmount must not be empty")
-        Double netAmount,
+        BigDecimal netAmount,
         @NotBlank(message = "receiptDate must not be blank")
         @ValidDateTime
         String receiptDate,
